pyproject.toml
[tool.poetry]
name = "mrmonitor"
version = "0.1.0"
description = "Monitor your MRs on Gitlab"
authors = ["Meesha <44530786+meesha7@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-gitlab = "^2.10.1"
python-dotenv = "^0.19.2"
rich = "^10.13.0"
arrow = "^1.2.1"
click = "^8.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
